Resueltos conflictos priorizando develop en todos los archivos Java.
Se conservan los archivos Docker/Kubernetes de esta rama.

+++ src/main/java/fe/banco_digital/service/MotorValidacionServiceImpl.java
@@ -53,7 +53,7 @@ public class MotorValidacionServiceImpl implements MotorValidacionService {
     @Override
     public ValidacionTransaccionResponseDTO validarCuentaParaDebito(Cuenta cuenta, BigDecimal monto) {
         if (cuenta == null) {
-            throw new CuentaNoEncontradaException(null);
+            throw new CuentaNoEncontradaException((Long) null);
         }
 
         if (cuenta.getEstado() == EstadoCuenta.BLOQUEADA) {
